+++ src/main/java/tk/hiddenname/smarthome/controller/device/DeviceSignalRestController.java
@@ -2,9 +2,9 @@ package tk.hiddenname.smarthome.controller.device;
 
 import lombok.AllArgsConstructor;
 import org.springframework.web.bind.annotation.*;
-import tk.hiddenname.smarthome.entity.hardware.Device;
-import tk.hiddenname.smarthome.entity.signal.DigitalState;
-import tk.hiddenname.smarthome.entity.signal.PwmSignal;
+import tk.hiddenname.smarthome.model.hardware.Device;
+import tk.hiddenname.smarthome.model.signal.DigitalState;
+import tk.hiddenname.smarthome.model.signal.PwmSignal;
 import tk.hiddenname.smarthome.service.database.DeviceDatabaseService;
 import tk.hiddenname.smarthome.service.hardware.impl.digital.output.DigitalDeviceServiceImpl;
 import tk.hiddenname.smarthome.service.hardware.impl.pwm.output.PwmDeviceServiceImpl;
@@ -23,7 +23,7 @@ public class DeviceSignalRestController {
 
     // ******************************** PWM *************************************************
     @GetMapping(value = {"/pwm"}, produces = {"application/json"})
-    public PwmSignal getPwmSignal(@RequestParam(name = "id") Integer id) {
+    public PwmSignal getPwmSignal(@RequestParam(name = "id") Long id) {
         Device device = dbService.getOne(id);
 
         return pwmService.getSignal(device.getId(), device.isReverse());
@@ -41,7 +41,7 @@ public class DeviceSignalRestController {
     // ***************************** DIGITAL **************************************************
 
     @GetMapping(value = {"/digital"}, produces = {"application/json"})
-    public DigitalState getState(@RequestParam(name = "id") Integer id) {
+    public DigitalState getState(@RequestParam(name = "id") Long id) {
         Device device = dbService.getOne(id);
 
         return digitalService.getState(device.getId(), device.isReverse());
