+++ src/main/java/tk/hiddenname/smarthome/exception/DeviceNotFoundException.java
@@ -6,7 +6,7 @@ import org.springframework.web.bind.annotation.ResponseStatus;
 @ResponseStatus(value = HttpStatus.NOT_FOUND)
 public class DeviceNotFoundException extends RuntimeException {
 
-    public DeviceNotFoundException(int id) {
+    public DeviceNotFoundException(Long id) {
         super(String.format("The device with id '%d' not found", id));
     }
 }
